Split term.rs into multiple files
